reduce sequence nr to 32 bits

+++ internal/gateway/packets.go
@@ -113,7 +113,7 @@ type evtIdentity struct {
 type evtResume struct {
 	Token      string `json:"token"`
 	SessionID  string `json:"session_id"`
-	SequenceNr uint64 `json:"seq"`
+	SequenceNr uint32 `json:"seq"`
 }
 
 type RequestGuildMembersPayload struct {
@@ -229,7 +229,7 @@ type helloPacket struct {
 type discordPacketJSON struct {
 	Op             opcode.OpCode `json:"op"`
 	Data           []byte        `json:"d"`
-	SequenceNumber uint64        `json:"s"`
+	SequenceNumber uint32        `json:"s"`
 	EventName      string        `json:"t"`
 }
 
@@ -244,7 +244,7 @@ func (p *discordPacketJSON) CopyOverTo(packet *DiscordPacket) {
 type DiscordPacket struct {
 	Op             opcode.OpCode   `json:"op"`
 	Data           json.RawMessage `json:"d"`
-	SequenceNumber uint64          `json:"s,omitempty"`
+	SequenceNumber uint32          `json:"s,omitempty"`
 	EventName      string          `json:"t,omitempty"`
 }
